requirements.txt
arrow
python-dotenv
typer
python-frontmatter
opensearch-py
langchain
sentence-transformers